Remove unecessary file
